Move interaction handling into hook
