+++ polaris-react/src/index.ts
@@ -437,3 +437,4 @@ export {
 } from './utilities/index-provider';
 export {useBreakpoints} from './utilities/breakpoints';
 export {useReadOnlyEphemeralPresenceManager as useEphemeralPresenceManager} from './utilities/ephemeral-presence-manager';
+export {useHoverCardActivatorWrapperProps} from './utilities/hover-card';
